requirements.txt
annotated-types==0.7.0
anyio==4.10.0
bottle==0.13.4
build==1.3.0
certifi==2025.8.3
cffi==1.17.1
click==8.2.1
clr_loader==0.2.7.post0
colorama==0.4.6
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
numpy==2.3.2
packaging==25.0
pillow==11.3.0
proxy_tools==0.1.0
pybind11==3.0.0
pybind11-global==3.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyproject_hooks==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
pythonnet==3.0.5
pywebview==6.0
PyYAML==6.0.2
rawpy==0.25.1
rich==14.1.0
rich-toolkit==0.15.0
rignore==0.6.4
scikit-build==0.18.1
sentry-sdk==2.35.0
setuptools==80.9.0
shellingham==1.5.4
sniffio==1.3.1
starlette==0.47.2
typer==0.16.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
wheel==0.45.1
